add dead queue check in SoraObject; fix SoraSpriteParser plugin;

+++ Plugins/SoraLua/LuaPlus/LuaPlusCD.h
@@ -1089,7 +1089,7 @@ namespace LPCD
 
 			Object* obj = (Object*)LPCD::GetObjectUserData(L);
 
-			LPCD::Push(L, *(VarType*)((unsigned char*)obj + (unsigned int)offset));
+			LPCD::Push(L, *(VarType*)((unsigned char*)obj + (unsigned long)offset));
 
 			return 1;
 		}
@@ -1103,7 +1103,7 @@ namespace LPCD
 			if (!Match(TypeWrapper<VarType>(), L, 2))
 				luaL_argerror(L, 2, "bad argument");
 
-			*(VarType*)((unsigned char*)obj + (unsigned int)offset) = Get(TypeWrapper<VarType>(), L, 2);
+			*(VarType*)((unsigned char*)obj + (unsigned long)offset) = Get(TypeWrapper<VarType>(), L, 2);
 
 			return 1;
 		}
